episode-09: Added custom hook for restaurant list
